requirements.txt
streamlit
boto3
python-dotenv
pandas
plotly
openpyxl
psycopg2-binary
sqlalchemy
langchain
langgraph
langchain-aws
langchain-mcp-adapters
langchain-core